Switch to event listener style which is immune to the opentelemetry http event listener remove bug
Somewhere deep inside the opentelemetry JS SDK, something breaks the request.removeEventListener such that adding the `abort` listener then removing it doesn't work. It still fires even after it has been removed! So broken, but, I haven't found the source of the bug in opentelemetry yet, so to move forward I want to not rely on removeListener working. Instead we can just set a boolean.